+++ fe/fe-core/src/main/java/com/starrocks/alter/SchemaChangeHandler.java
@@ -36,9 +36,9 @@ import com.starrocks.analysis.AlterClause;
 import com.starrocks.analysis.CancelAlterTableStmt;
 import com.starrocks.analysis.CancelStmt;
 import com.starrocks.analysis.ColumnPosition;
-import com.starrocks.sql.ast.CreateIndexClause;
+import com.starrocks.analysis.CreateIndexClause;
 import com.starrocks.analysis.DropColumnClause;
-import com.starrocks.sql.ast.DropIndexClause;
+import com.starrocks.analysis.DropIndexClause;
 import com.starrocks.analysis.IndexDef;
 import com.starrocks.analysis.ModifyColumnClause;
 import com.starrocks.analysis.ModifyTablePropertiesClause;
@@ -1850,6 +1850,10 @@ public class SchemaChangeHandler extends AlterHandler {
     private void processDropIndex(DropIndexClause alterClause, OlapTable olapTable, List<Index> indexes)
             throws DdlException {
         String indexName = alterClause.getIndexName();
+        //Mock syntax Alter table tbl drop primary key;
+        if (indexName.equals("PRIMARY")) {
+            return;
+        }
         List<Index> existedIndexes = olapTable.getIndexes();
         Index found = null;
         for (Index existedIdx : existedIndexes) {
