--- fe/fe-core/src/test/java/com/starrocks/analysis/DropIndexClauseTest.java
@@ -1,52 +1,0 @@
-// This file is made available under Elastic License 2.0.
-// This file is based on code available under the Apache license here:
-//   https://github.com/apache/incubator-doris/blob/master/fe/fe-core/src/test/java/org/apache/doris/analysis/DropIndexClauseTest.java
-
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-package com.starrocks.analysis;
-
-import com.starrocks.common.AnalysisException;
-import com.starrocks.common.UserException;
-import com.starrocks.sql.ast.DropIndexClause;
-import org.junit.Assert;
-import org.junit.BeforeClass;
-import org.junit.Test;
-
-public class DropIndexClauseTest {
-
-    private static Analyzer analyzer;
-
-    @BeforeClass
-    public static void setUp() {
-        analyzer = AccessTestUtil.fetchAdminAnalyzer(false);
-    }
-
-    @Test
-    public void testNormal() throws UserException {
-        DropIndexClause clause = new DropIndexClause("index1", new TableName("db", "table"), false);
-        clause.analyze(analyzer);
-        Assert.assertEquals("DROP INDEX index1 ON `db`.`table`", clause.toSql());
-    }
-
-    @Test(expected = AnalysisException.class)
-    public void testNoIndex() throws UserException {
-        DropIndexClause clause = new DropIndexClause("", new TableName("db", "table"), false);
-        clause.analyze(analyzer);
-    }
-}
